// src/app/features/aspd/types/equip.ts
export enum Equip {
  BareHand = 'Bare Hand',
  Shield = 'Shield',
  Dagger = 'Dagger',
  DaggerLeft = 'Dagger (Left)',
  Bow = 'Bow',
  Instrument = 'Instrument',
  Whip = 'Whip',
  Mace = 'Mace',
  OneHandSword = 'One-Hand Sword',
  OneHandSwordLeft = 'One-Hand Sword (Left)',
  TwoHandSword = 'Two-Hand Sword',
  OneHandAxe = 'One-Hand Axe',
  OneHandAxeLeft = 'One-Hand Axe (Left)',
  TwoHandAxe = 'Two-Hand Axe',
  OneHandSpear = 'One-Hand Spear',
  TwoHandSpear = 'Two-Hand Spear',
  OneHandStaff = 'One-Hand Staff',
  TwoHandStaff = 'Two-Hand Staff',
  Book = 'Book',
  Knuckle = 'Knuckle',
  Katar = 'Katar',
}