// src/app/features/aspd/types/class.ts
export enum Class {
  Novice = 'Novice',

  Archer = 'Archer',
  Hunter = 'Hunter',
  Bard = 'Bard',
  Dancer = 'Dancer',

  Swordman = 'Swordman',
  Knight = 'Knight',
  Crusader = 'Crusader',

  Merchant = 'Merchant',
  Blacksmith = 'Blacksmith',
  Alchemist = 'Alchemist',

  Mage = 'Mage',
  Wizard = 'Wizard',
  Sage = 'Sage',

  Acolyte = 'Acolyte',
  Priest = 'Priest',
  Monk = 'Monk',

  Thief = 'Thief',
  Assassin = 'Assassin',
  Rogue = 'Rogue',
}